# AvroType performace improvements