This PR aims to improve the derived typeclass performance for Avro

Results of `jmh` benchmarks with `jmh/jmh:run -i 20 -wi 10 -f1 -t1 magnolify.jmh.AvroBench` are ~actually not~ looking good

| branch | `avroTo` | `avroFrom` | `avroSchema` |
|-----------|------------|-----------------|---------------------|
| ~PR~       |  ~8981.511 ns/op~  | ~6675.408 ns/op~ | ~3.697 ns/op~ |
| PR       |  3967.037 ns/op  | 1718.793 ns/op | 4.016 ns/op |
| main    |  8075.481 ns/op  | 3190.532 ns/op | 3.832 ns/op |
| v0.4.8  |  9463.751 ns/op  | 3417.335 ns/op | 3.767 ns/op |

Looks like `rawConstruct` is making an extra call to `MagnoliaUtil.checkParamLengths` wich is worse that going though the field map in the schema

Edit: Using vars instead improves greatly avro conversion performances (~ 40% to 50% compared to v0.4.8)