@kellen 2022-07-05
Seems like codecov would be made happy by testing `DateTimeLogicalTypeFactory`