@RustedBones 2022-07-06
Strange. It should actually be tested [here](https://github.com/spotify/magnolify/blob/c4d34c8977c078c2ecfdd62980e36929f4fa2dce/avro/src/test/scala/magnolify/avro/test/AvroTypeSuite.scala#L198)